# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Article API",
    "description": "Article API to manage Admin users, Tenants, Articles, Tags and References.",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/api/",
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/articles": {
      "post": {
        "tags": ["Article"],
        "description": "Create new Article",
        "operationId": "createArticle",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Article Parameters",
            "in": "body",
            "description": "Create new Article",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Article"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "403": {
            "description": "Invalid Parameter",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },      
      "get": {
        "tags": ["Article"],
        "description": "Get Article List",
        "operationId": "getArticles",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Article List",
            "schema": {
              "$ref": "#/definitions/GetArticle"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/articles/{_id}": {
      "patch": {
        "tags": ["Article"],
        "description": "Update Article",
        "operationId": "updateArticle",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Article Parameters",
            "in": "body",
            "description": "Update Article",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateArticle"
            }
          },
          {
            "name": "_id",
            "in": "path",
            "description": "Article Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Invalid Article Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": ["Article"],
        "description": "Delete Article",
        "operationId": "deleteArticle",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          },
          {
            "name": "_id",
            "in": "path",
            "description": "_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "404": {
            "description": "Invalid Article Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateArticle": {
      "properties": {
        "title": {
          "required":true,
          "type": "string",
          "example": "Title"
        },
        "description": {
          "required":true,
          "type": "string",
          "example": "Description"
        },
        "type": {
          "required":true,
          "type": "string",
          "example": "Type"
        },
        "image": {
          "required":true,
          "type": "string",
          "example": "Image"
        },
        "start_date": {
          "required":true,
          "type": "string",
          "example": "2012/4/4"
        },
        "expiry_date": {
          "required":true,
          "type": "string",
          "example": "2016/5/5"
        },
        "token": {
          "required":true,
          "type": "string",
          "example": "Token"
        }
      }
    },
    "GetArticle": {
      "properties": {
        "title": {
          "required":true,
          "type": "string",
          "example": "Title"
        },
        "description": {
          "required":true,
          "type": "string",
          "example": "Description"
        },
        "image": {
          "required":true,
          "type": "string",
          "example": "Image"
        },
        "type": {
          "required":true,
          "type": "string",
          "example": "Type"
        },
        "start_date": {
          "required":true,
          "type": "string",
          "example": "2012/4/4"
        },
        "expiry_date": {
          "required":true,
          "type": "string",
          "example": "2016/5/5"
        }
      }
    },
    "Article": {
      "properties": {
        "title": {
          "required":true,
          "type": "string",
          "example": "Title"
        },
        "description": {
          "required":true,
          "type": "string",
          "example": "Description"
        },
        "type": {
          "required":true,
          "type": "string",
          "example": "Type"
        },
        "image": {
          "required":true,
          "type": "string",
          "example": "Image"
        },
        "start_date": {
          "required":true,
          "type": "string",
          "example": "2012/4/4"
        },
        "expiry_date": {
          "required":true,
          "type": "string",
          "example": "2016/5/5"
        },
        "token": {
          "required":true,
          "type":"string",
          "example": "Token"
        } 
      }
    },
...
  }
}